Fix circular import issue when running in repl or from cli
