+++ backend/doc/domain/resource_lookup.py
@@ -16,7 +16,7 @@ from typing import Mapping, Optional, Sequence
 import requests
 from cachetools import TTLCache, cached
 from doc.config import settings
-from doc.domain import worker, parsing
+from doc.domain import worker
 from doc.domain.bible_books import BOOK_CHAPTERS, BOOK_ID_MAP, BOOK_NAMES
 from doc.domain.model import (
     NON_USFM_RESOURCE_TYPES,
@@ -263,6 +263,8 @@ def get_resource_types(
         str, str
     ] = settings.RESOURCE_TYPE_CODES_AND_NAMES,
 ) -> list[tuple[str, str]]:
+    from doc.domain.parsing import find_usfm_files
+
     resource_types = []
     for url, resource_filepath, resource_type in repo_clone_list:
         if resource_type:
@@ -285,7 +287,7 @@ def get_resource_types(
                     and file.name.split("-")[1].lower() in book_codes
                 ]
             elif resource_type in usfm_resource_types:
-                book_assets = parsing.find_usfm_files(resource_filepath)
+                book_assets = find_usfm_files(resource_filepath)
             elif resource_type == "rg":
                 between_texts, bible_reference_strs = find_bible_references(
                     join(en_rg, docx_file_path)
@@ -574,6 +576,8 @@ def usfm_resource_types_and_book_tuples(
     >>> sorted(tuples, key=lambda value: value[1])
     [('reg', '1co'), ('reg', '1jn'), ('reg', '1pe'), ('reg', '1th'), ('reg', '1ti'), ('reg', '2co'), ('reg', '2jn'), ('reg', '2pe'), ('reg', '2th'), ('reg', '2ti'), ('reg', '3jn'), ('reg', 'act'), ('reg', 'col'), ('reg', 'eph'), ('reg', 'gal'), ('reg', 'heb'), ('reg', 'jas'), ('reg', 'jhn'), ('reg', 'jud'), ('reg', 'luk'), ('reg', 'mat'), ('reg', 'mrk'), ('reg', 'phm'), ('reg', 'php'), ('reg', 'rev'), ('reg', 'rom'), ('reg', 'tit')]
     """
+    from doc.domain.parsing import usfm_asset_file
+
     book_codes = book_codes_str.split(",")
     data: SourceData | None = fetch_source_data()
     resource_type_and_book_tuples = set()
@@ -602,9 +606,7 @@ def usfm_resource_types_and_book_tuples(
                     resource_filepath = prepare_resource_filepath(dto)
                     if file_needs_update(resource_filepath):
                         provision_asset_files(dto.url, resource_filepath)
-                    content_file = parsing.usfm_asset_file(
-                        dto, resource_filepath, False
-                    )
+                    content_file = usfm_asset_file(dto, resource_filepath, False)
                     if content_file:
                         resource_type_and_book_tuples.add((resource_type, book_code))
     return sorted(resource_type_and_book_tuples, key=lambda value: value[0])
@@ -942,18 +944,24 @@ def get_book_names_from_usfm_metadata(
     be localized, it depends on the translation work done for language
     lang_code.
     """
+    from doc.domain.parsing import (
+        find_usfm_files,
+        split_usfm_by_chapters,
+        maybe_localized_book_name,
+    )
+
     book_codes_and_names_localized: dict[str, str] = {}
-    usfm_files = parsing.find_usfm_files(resource_filepath)
+    usfm_files = find_usfm_files(resource_filepath)
     for usfm_file in usfm_files:
         usfm = ""
         usfm_file_components = Path(usfm_file).stem.lower().split("-")
         book_code = usfm_file_components[1]
         with open(usfm_file, "r") as f:
             usfm = f.read()
-        frontmatter, _, _ = parsing.split_usfm_by_chapters(
+        frontmatter, _, _ = split_usfm_by_chapters(
             lang_code, resource_type, book_code, usfm
         )
-        localized_book_name = parsing.maybe_localized_book_name(frontmatter)
+        localized_book_name = maybe_localized_book_name(frontmatter)
         # localized_book_name = maybe_correct_book_name(lang_code, localized_book_name)
         book_codes_and_names_localized[book_code] = localized_book_name
     logger.debug("book_codes_and_names_localized: %s", book_codes_and_names_localized)
